Add: Testapp SSL cert and updated resource usage
It turns out that there are cases when we might want to create
but not enable, or disable but not delete a vhost.  This test
cookbook uses some of these actions.

The default vhost needs to be removed because it contains an SSL
directive that conflicts with the one our LWRP generates.